\name{print.nbdprogram}
\alias{print.nbdprogram}
\title{Print results}
\usage{
\method{print}{nbdprogram}(x, ...)
}
\arguments{
\item{x}{A list object of class \code{nbdprogram}.}
\item{...}{Other parameters passing to the generic function}
}
\description{
Print results of \link{nbdprogram}.
}
\seealso{
\code{\link{nbdprogram}}
}
